fix gdal=3.9 bug

+++ pywapor/general/curvilinear.py
@@ -86,7 +86,7 @@ def curvi_to_recto(lats, lons, data, out_fn, warp_kwargs = {}):
 
 def create_grid(latlim, lonlim, dx_dy = (0.0033, 0.0033)):
     dx, dy = dx_dy
-    nx = np.ceil((lonlim[1] - lonlim[0]) / dx)
-    ny = np.ceil((latlim[1] - latlim[0]) / dy)
+    nx = int(np.ceil((lonlim[1] - lonlim[0]) / dx))
+    ny = int(np.ceil((latlim[1] - latlim[0]) / dy))
     bb = [lonlim[0], latlim[0], lonlim[0] + nx * dx, latlim[0] + ny * dy]
     return bb, nx, ny
